# Source_code/RestAPI/Definition/abdarum-libraryVarna-1.0.1-oas3-resolved.yaml
...
openapi: "3.0.0"
info:
  title: "Library Varna"
  description: "Library API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "kornelstefanczyk@wp.pl"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  version: "1.0.1-oas3"
...
servers:
- url: "http://localhost:8080"
security:
- bearerAuth: []
...
paths:
...
  /book:
    put:
      tags:
      - "book"
      summary: "Update an existing book"
      operationId: "updateBook"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Book not found"
        "405":
          description: "Validation exception"
      security:
      - bearerAuth: []
    post:
      tags:
      - "book"
      summary: "Add a new book to the library"
      operationId: "addBook"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        "405":
          description: "Invalid input"
      security:
      - bearerAuth: []
  /book/{id}:
    get:
      tags:
      - "book"
      summary: "Get book by book id"
      operationId: "getBookById"
      parameters:
      - name: "id"
        in: "path"
        description: "The id that needs to be fetched."
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Book not found"
    put:
      tags:
      - "book"
      summary: "Updated book"
      description: "This can only be done by the logged in user."
      operationId: "updateBookById"
      parameters:
      - name: "id"
        in: "path"
        description: "id of book that need to be updated"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      requestBody:
        description: "Updated user object"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Book not found"
    delete:
      tags:
      - "book"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteBookById"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the book that needs to be deleted"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "400":
          description: "Invalid id supplied"
        "404":
          description: "Book not found"
  /book/findByStatus:
    get:
      tags:
      - "book"
      summary: "Finds Books by status"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findBookByStatus"
      parameters:
      - name: "status"
        in: "query"
        description: "Status values that need to be considered for filter"
        required: true
        style: "form"
        explode: true
        schema:
          type: "array"
          items:
            type: "string"
            enum:
            - "available"
            - "user"
            - "name"
            - "author"
            default: "available"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Book"
        "400":
          description: "Invalid status value"
      security:
      - bearerAuth: []
  /library/inventory:
    get:
      tags:
      - "library"
      summary: "Returns library inventories by status"
      description: "Returns a stats of the library inventory"
      operationId: "getLibraryInventory"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryStats"
      security:
      - bearerAuth: []
  /library/order:
    post:
      tags:
      - "library"
      summary: "Place an order for a book rent"
      operationId: "placeOrder"
      requestBody:
        description: "order placed for borrow the book"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Borrowed"
        required: true
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowed"
        "400":
          description: "Invalid Order"
  /library/order/{orderId}:
    get:
      tags:
      - "library"
      summary: "Find book order by ID"
      operationId: "getOrderById"
      parameters:
      - name: "orderId"
        in: "path"
        description: "ID of borrow form that needs to be fetched"
        required: true
        style: "simple"
        explode: false
        schema:
          maximum: 1E+1
          minimum: 1
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowed"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"
    put:
      tags:
      - "library"
      summary: "Find purchase order by ID"
      operationId: "putOrderById"
      parameters:
      - name: "orderId"
        in: "path"
        description: "ID of pet that needs to be fetched"
        required: true
        style: "simple"
        explode: false
        schema:
          maximum: 1E+1
          minimum: 1
          type: "integer"
          format: "int64"
      requestBody:
        description: "form ot the borrowed book that needs to update"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Borrowed"
        required: true
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowed"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"
    delete:
      tags:
      - "library"
      summary: "Delete purchase order by ID"
      operationId: "deleteOrder"
      parameters:
      - name: "orderId"
        in: "path"
        description: "ID of the order that needs to be deleted"
        required: true
        style: "simple"
        explode: false
        schema:
          minimum: 1
          type: "integer"
          format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Order not found"
...
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        userTypeId:
          type: "integer"
          format: "int64"
        firstName:
          type: "string"
        lastName:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        phone:
          type: "string"
        birthdate:
          type: "integer"
          description: "Timestamp in UTC"
        registrated:
          type: "integer"
          description: "Timestamp in UTC"
        gender:
          type: "string"
        adress:
          type: "string"
        city:
          type: "string"
        status:
          type: "string"
          description: "Order Status"
          enum:
          - "active"
          - "suspended"
          - "inactive"
      xml:
        name: "User"
...
    Borrowed:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        userId:
          type: "integer"
          format: "int64"
        bookId:
          type: "integer"
          format: "int64"
        borrowedDate:
          type: "integer"
          description: "Timestamp in UTC"
        returnedDate:
          type: "integer"
          description: "Timestamp in UTC"
        damageNotes:
          type: "string"
        placeId:
          type: "integer"
          format: "int64"
        periodId:
          type: "integer"
          format: "int64"
      xml:
        name: "Borrowed"
    Book:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        isbn:
          type: "integer"
        authorId:
          type: "integer"
          format: "int64"
        pageCount:
          type: "integer"
        coverTypeId:
          type: "integer"
          format: "int64"
        genreId:
          type: "integer"
          format: "int64"
      xml:
        name: "Book"
...
    LibraryStats:
      type: "object"
      properties:
        numberOfBooks:
          type: "integer"
        numberOfAvailableBooks:
          type: "integer"
      xml:
        name: "LibraryStats"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"